backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.23
psycopg[binary]==3.2.1
pydantic==2.8.2 
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
alembic==1.13.1
pytest==7.4.3
alembic==1.13.1
docstrange==1.1.8
openai==1.40.5
python.dotenv==1.0.1
httpx==0.27.0
requests==2.32.3
stripe==8.6.0
plaid-python==15.0.0
pandas>=2.0.0
openpyxl>=3.1.0